+++ example/lib/bounded_page.dart
@@ -1,45 +1,46 @@
 import 'dart:math' as math;
 
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 
 import 'package:mobile_window_features/mobile_window_features.dart';
+import 'package:mobile_window_features/status_navigation_bars_options.dart';
 
 class BoundedPage extends StatelessWidget {
   const BoundedPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      decoration: createBackground(),
-      child: Column(children: const [Text("pruebas")]),
+    MobileWindowFeatures.setScreenProperties(
+        ScreenLimits.pageDrawLimitedByStatusBarNavigationBar,
+        Colors.blue,
+        StatusBarThemeMWF.darkStatusBar,
+        Colors.white,
+        NavigationBarThemeMWF.lightNavigationBar);
+
+    return Scaffold(
+      body: Stack(
+        children: [
+          Container(
+            decoration: createBackground(),
+          ),
+          Column(children: [
+            Row(
+              children: const [
+                Expanded(
+                    child: Text(
+                  "Test",
+                  style: TextStyle(color: Colors.white, fontSize: 40),
+                  textAlign: TextAlign.center,
+                ))
+              ],
+            ),
+          ]),
+        ],
+      ),
     );
   }
 
   BoxDecoration createBackground() {
-    MobileWindowFeatures.clearFlags(MobileWindowFeatures.FLAG_LAYOUT_NO_LIMITS);
-    MobileWindowFeatures.setFlags(
-        MobileWindowFeatures.FLAG_LAYOUT_IN_SCREEN |
-            MobileWindowFeatures.FLAG_LAYOUT_INSET_DECOR,
-        MobileWindowFeatures.FLAG_LAYOUT_IN_SCREEN |
-            MobileWindowFeatures.FLAG_LAYOUT_INSET_DECOR);
-    MobileWindowFeatures.addFlags(
-        MobileWindowFeatures.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
-    MobileWindowFeatures.clearFlags(
-        MobileWindowFeatures.FLAG_TRANSLUCENT_STATUS);
-    MobileWindowFeatures.clearFlags(
-        MobileWindowFeatures.FLAG_TRANSLUCENT_NAVIGATION);
-    MobileWindowFeatures.setDecorFitsSystemWindows(true);
-
-    SystemChrome.setSystemUIOverlayStyle(const SystemUiOverlayStyle(
-      statusBarColor: Colors.transparent,
-      statusBarIconBrightness: Brightness.dark,
-      systemNavigationBarColor: Colors.transparent,
-      systemNavigationBarDividerColor: Colors.transparent,
-      systemStatusBarContrastEnforced: true,
-      systemNavigationBarContrastEnforced: true,
-    ));
-
     return const BoxDecoration(
         gradient: LinearGradient(
             begin: Alignment.topCenter,