+++ lib/mobile_window_features_method_channel.dart
@@ -7,7 +7,8 @@ import 'mobile_window_features_platform_interface.dart';
 class MethodChannelMobileWindowFeatures extends MobileWindowFeaturesPlatform {
   /// The method channel used to interact with the native platform.
   @visibleForTesting
-  final methodChannel = const MethodChannel('mobile_window_features');
+  final MethodChannel methodChannel =
+      const MethodChannel('mobile_window_features');
 
   /// Adds flags [flags] to the WindowManager.LayoutParams
   @override
@@ -39,6 +40,41 @@ class MethodChannelMobileWindowFeatures extends MobileWindowFeaturesPlatform {
   Future<bool?> setDecorFitsSystemWindows(bool decorFitsSystemWindows) async {
     return await methodChannel.invokeMethod<bool>("setDecorFitsSystemWindows", {
       "decorFitsSystemWindows": decorFitsSystemWindows,
+    });
+  }
+
+  @override
+  Future<bool?> setScreenLimits(int screenLimits) async {
+    return await methodChannel.invokeMethod<bool>("setScreenLimits", {
+      "screenLimits": screenLimits,
+    });
+  }
+
+  @override
+  Future<bool?> setStatusBarColor(int color) async {
+    return await methodChannel.invokeMethod<bool>("setStatusBarColor", {
+      "colorInt": color,
+    });
+  }
+
+  @override
+  Future<bool?> setNavigationBarColor(int color) async {
+    return await methodChannel.invokeMethod<bool>("setNavigationBarColor", {
+      "colorInt": color,
+    });
+  }
+
+  @override
+  Future<bool?> setStatusBarTheme(int statusBarTheme) async {
+    return await methodChannel.invokeMethod<bool>("setStatusBarTheme", {
+      "statusBarTheme": statusBarTheme,
+    });
+  }
+
+  @override
+  Future<bool?> setNavigationBarTheme(int navigationBarTheme) async {
+    return await methodChannel.invokeMethod<bool>("setNavigationBarTheme", {
+      "navigationBarTheme": navigationBarTheme,
     });
   }
 }