+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-## 0.0.1
+## 0.0.2
+- Allow to change the color of the status bar
+- Allow to change the color of the navigation bar
+- Allow to change the color of the icons in the status bar
+- Allow to change the color of the icons in the navigation bar
+- Allow to draw behind the status bar and navigation bar using only one method
+- Allow to draw the status bar and navigation bar with an specific color using only one method
 
-* TODO: Describe initial release.
+## 0.0.1
+- Allow to "addFlags" on the system window of android
+- Allow to "setFlags" on the system window of android
+- Allow to "clearFlags" on the system window of android
+- Allow to "setDecorFitsSystemWindows" on the system window of android
+
+* Allow to change some properties on the system window of android